+++ src/main/java/de/muspellheim/commons/fx/control/StatusBarSkin.java
@@ -25,11 +25,11 @@ class StatusBarSkin extends SkinBase<StatusBar> {
         getChildren().add(root);
 
         HBox leftItems = new HBox();
+        leftItems.setAlignment(Pos.CENTER_LEFT);
         leftItems.setSpacing(5);
         leftItems.getStyleClass().add("left-items");
         leftItems.getChildren().setAll(getSkinnable().getLeftItems());
         GridPane.setMargin(leftItems, new Insets(0, 10, 0, 0));
-        GridPane.setFillHeight(leftItems, true);
         GridPane.setVgrow(leftItems, Priority.ALWAYS);
         root.add(leftItems, 0, 0);
 
@@ -37,18 +37,21 @@ class StatusBarSkin extends SkinBase<StatusBar> {
         text.setMaxSize(Double.MAX_VALUE, Double.MAX_VALUE);
         text.getStyleClass().add("status-text");
         GridPane.setHgrow(text, Priority.ALWAYS);
+        GridPane.setVgrow(text, Priority.ALWAYS);
         root.add(text, 1, 0);
 
         ProgressBar progress = new ProgressBar();
         progress.getStyleClass().add("status-progress");
+        GridPane.setMargin(progress, new Insets(0, 0, 0, 5));
+        GridPane.setVgrow(progress, Priority.ALWAYS);
         root.add(progress, 2, 0);
 
         HBox rightItems = new HBox();
+        rightItems.setAlignment(Pos.CENTER_RIGHT);
         rightItems.setSpacing(5);
         rightItems.getStyleClass().add("right-items");
         rightItems.getChildren().setAll(getSkinnable().getRightItems());
         GridPane.setMargin(rightItems, new Insets(0, 0, 0, 10));
-        GridPane.setFillHeight(rightItems, true);
         GridPane.setVgrow(rightItems, Priority.ALWAYS);
         root.add(rightItems, 3, 0);
 
@@ -65,7 +68,6 @@ class StatusBarSkin extends SkinBase<StatusBar> {
         BooleanBinding notZeroProgress = Bindings.notEqual(0, getSkinnable().progressProperty());
         progress.progressProperty().bind(getSkinnable().progressProperty());
         progress.visibleProperty().bind(notZeroProgress);
-        progress.managedProperty().bind(notZeroProgress);
 
         rightItems.managedProperty().bind(Bindings.isNotEmpty(getSkinnable().getRightItems()));
         getSkinnable().getRightItems().addListener((Observable o) -> rightItems.getChildren().setAll(getSkinnable().getRightItems()));