+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+*   Toggle process changed status bar height. 
+
 ## [v3.2.0] - 2019-11-05
 
 ### Added